Avoid using the BlockingProcess dispatcher for IO
The BlockingProcess actor's dispatcher was leveraged for use by various IO operations within it. For some undiscovered reason, this limited the BlockingProcess to become blocked at around the minimum number of threads per the core setting of its ThreadPoolExecutor.

I was able to create a new test that started up 100 processes, and therefore 300 threads on a dedicated dispatcher. Without the changes here i.e. with the BlockingProcess actor sharing its TPE with the stdio stream sources and sinks, the test often failed and hung. With the changes where the TPE is dedicated to the IO, and the BlockingProcess actor resides on the default dispatcher, I was unable to reproduce the issue.

In addition, having a dispatcher dedicated to the actor's IO seems more focused.

Finally, the problem of seeing that threads were blocked in the manner described here was observed at a ConductR customer's site. By applying these changes I'm optimistic that that particular issue will be resolved.